backend/requirements.txt
Django>=6.0
djangorestframework>=3.16
django-cors-headers>=4.0
Pillow>=12.0
anthropic>=0.25
python-dotenv>=1.0
pillow-heif>=0.18
